requirements.txt
# Dipendenze principali per Minecraft Skin AI Generator
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0

# Dipendenze opzionali per il training avanzato
matplotlib>=3.5.0
scikit-learn>=1.0.0

# Per sistemi Windows con CUDA (opzionale)
# torch --index-url https://download.pytorch.org/whl/cu118

# Dipendenze di sviluppo (opzionali)
jupyter>=1.0.0
tensorboard>=2.8.0 